:white_check_mark: chore: update public package
