+++ .gitignore
@@ -2,6 +2,8 @@
 node_modules
 *.log
 bin
+esm
+lib
 .eslintcache
 coverage
 yarn.lock
