+++ package.json
@@ -33,6 +33,7 @@
     "@attachments/serendipity-workflows": "latest",
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
+    "@lint-md/eslint-plugin": "^0.0.3",
     "@types/fs-extra": "^9.0.7",
     "@types/jest": "^26.0.20",
     "@types/node": "^14.14.22",
@@ -44,13 +45,14 @@
     "eslint": "^7.18.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-import": "^2.22.1",
-    "@lint-md/eslint-plugin": "^0.0.3",
     "eslint-plugin-prettier": "^3.3.1",
     "execa": "^5.0.0",
     "fs-extra": "^9.1.0",
     "husky": "^4.3.8",
     "jest": "^26.6.3",
     "lerna": "^3.22.1",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "^3.0.2",
     "ts-jest": "^26.5.2",
     "typescript": "^4.2.2"
   },
