+++ packages/serendipity-core/__tests__/presetManager.spec.ts
@@ -8,9 +8,9 @@
 
 import MockAdapter from 'axios-mock-adapter'
 import axios from 'axios'
-import { generateTempPathInfo } from '../src'
-import PresetManager from '../src/utils/presetManager'
-import { PRESET_CDN_BASE_URL } from '../src/common/constant'
+import { generateTempPathInfo } from '@attachments/serendipity-public'
+import { PRESET_CDN_BASE_URL } from '@attachments/serendipity-public/bin/common/constant'
+import PresetManager from '../src/presetManager'
 
 // eslint-disable-next-line max-lines-per-function
 describe('preset Manager 模块测试', () => {
