+++ packages/serendipity-core/src/common/pluginMetaKeys.ts
