+++ packages/serendipity-core/src/coreManager.ts
@@ -9,12 +9,13 @@
 
 import * as fs from 'fs'
 import * as path from 'path'
-import { isPlugin, logger, PresetManager } from '@attachments/serendipity-public'
+import { isPlugin, logger } from '@attachments/serendipity-public'
 import { CreateOptions } from '@attachments/serendipity-public/bin/types/common'
 import { SerendipityPreset } from '@attachments/serendipity-public/bin/types/preset'
 import ConstructionManager from './constructionManager'
 import { AddOption } from './types/options'
 import createCoreManagerHooks from './hooks/coreManagerHooks'
+import PresetManager from './presetManager'
 
 
 class CoreManager {
