+++ packages/serendipity-core/src/pluginMeta.ts
