+++ packages/serendipity-core/src/presetManager.ts
@@ -9,10 +9,9 @@
 import * as path from 'path'
 import * as fs from 'fs'
 import axios from 'axios'
-import { SerendipityPreset } from '../types/preset'
-import { DEFAULT_PRESET_NAME, DEFAULT_PROJECT_NAME, PRESET_CDN_BASE_URL } from '../common/constant'
-import { writeFilePromise } from './files'
-import logger from './logger'
+import { logger, writeFilePromise } from '@attachments/serendipity-public'
+import { SerendipityPreset } from './types/preset'
+import { DEFAULT_PRESET_NAME, DEFAULT_PROJECT_NAME, PRESET_CDN_BASE_URL } from './common/constant'
 
 class PresetManager {
   private readonly basePath: string
