--- packages/serendipity-public/src/types/cliService.ts
@@ -1,13 +1,0 @@
-/*
- * File: cliService.ts
- * Description: service 层公共类型声明
- * Created: 2021-1-28 22:13:44
- * Author: yuzhanglong
- * Email: yuzl1123@163.com
- */
-
-// 合并 package.json 配置选项
-export interface MergePackageConfigOptions {
-  merge?: boolean
-  ignoreNullOrUndefined?: boolean
-}
